Support Multiple Header Sections In Configuration
This allows for multiple `license` configurations to be contained
within the same .licenserc.yaml file. The intended use is for a project
that has multiple packages or modules with differing licenses.

To do this we introduce a ConfigV2 config format since it is not valid
YAML for a key to be either a dictionary or a sequence.

We try to parse as V2 first, and fall back to V1.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>